// my-app/src/app/modelo/registro-paciente-dto.ts
export class RegistroPacienteDTO {
    cedula: string = "";
    nombre: string = "";
    celular: string = "";
    urlFoto: string = "";
    ciudad: string = "";
    fechaNacimiento: string = "";
    alergias: string = "";
    eps: string = "";
    tipoSangre: string = "";
    correo: string = "";
    confirmaPassword: string = "";
    contrasena: string = "";
}